pyproject.toml
[tool.poetry]
name = "referral-system-django"
version = "0.1.0"
description = ""
authors = ["bulatsabirow <bulatsabirow@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
django-rest-framework = "^0.1.0"
psycopg2-binary = "^2.9.9"
drfpasswordless = "^1.5.9"
django-phonenumber-field = { extras = ["phonenumbers"], version = "^7.3.0" }
drf-spectacular = "^0.27.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
